Change output to only write to file on success
Before this change, output behavior streams writes to file during the
mirror testing process. Interrupting this process results in a corrupt
or empty saved mirrorlist file. To avoid this failure mode, this changes
the default behavior to either write the entire output of a successful
run to the `--save` file path, or not at all.