Format speed test elapsed and connection time debug output
The current debug output for speed test results is needlessly precise,
with 5 or more decimals after both milliseconds and seconds.

Instead of printing this precision, implement a generic duration
formatter in the debugger for SpeedTestResult and make formatter
functions for both `elapsed` and `connection_time`.

In the generic duration formatter, format milliseconds without decimals
and use two decimals for seconds.

The debug output is much cleaner with the new snippet below.

  [CH] SpeedTestResult { speed: 257.6 KB/s; elapsed: 992ms; connection_time: 598ms }
  [NZ] SpeedTestResult { speed: 459.5 KB/s; elapsed: 962ms; connection_time: 1.02s }
  [NZ] SpeedTestResult { speed: 311.9 KB/s; elapsed: 995ms; connection_time: 1.02s }
  [SG] SpeedTestResult { speed: 238.0 KB/s; elapsed: 962ms; connection_time: 1.21s }
      TOP NEIGHBOR - CONNECTION TIME: MX - 108ms
      TOP NEIGHBOR - SPEED: SE - 56.9 MB/s